[MIG] product_code_builder from 8.0
