+++ .travis.yml
@@ -26,8 +26,10 @@ env:
   matrix:
   - LINT_CHECK="1"
   - TRANSIFEX="1"
-  - TESTS="1" ODOO_REPO="odoo/odoo"
-  - TESTS="1" ODOO_REPO="OCA/OCB"
+  - TESTS="1" EXCLUDE="product_sequence" ODOO_REPO="odoo/odoo"
+  - TESTS="1" INCLUDE="product_sequence" ODOO_REPO="odoo/odoo"
+  - TESTS="1" EXCLUDE="product_sequence" ODOO_REPO="OCA/OCB"
+  - TESTS="1" INCLUDE="product_sequence" ODOO_REPO="OCA/OCB"
 
 virtualenv:
   system_site_packages: true
