+++ product_code_builder/README.rst
@@ -23,7 +23,7 @@ Attributes order is preserved while code creation.
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/135/8.0
+   :target: https://runbot.odoo-community.org/runbot/135/9.0
 
 |
 
@@ -56,6 +56,12 @@ Warning
 
 This module may be not compatible with other modules which manage default_code like product_sequence
 
+Usage
+=====
+
+.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
+   :alt: Try me on Runbot
+   :target: https://runbot.odoo-community.org/runbot/135/9.0
 
 Bug Tracker
 ===========
@@ -66,15 +72,7 @@ check there if your issue has already been reported. If you spotted it first,
 help us smashing it by providing a detailed and welcomed `feedback
 <https://github.com/OCA/product-attribute/issues/new?body=module:%20
 product_code_builder%0Aversion:%20
-8.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-Known issues / Roadmap
-======================
-
-MIGRATION INSTRUCTION: Some fields in this module are the same than those in 
-product_attribute_priority module. For v10 or v9, it's required to add 
-dependency between these modules.
-
+9.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Credits
 =======
