+++ product_code_builder/models/__init__.py
@@ -1,3 +1,5 @@
 
 from . import product_attribute
-from . import product
+from . import product_attribute_value
+from . import product_template
+from . import product_product
