+++ product_code_builder/models/product_attribute_value.py
@@ -1,0 +1,18 @@
+# coding: utf-8
+# © 2015 Benoît GUILLOT <benoit.guillot@akretion.com>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from openerp import fields, models
+
+
+class ProductAttributeValue(models.Model):
+    _inherit = "product.attribute.value"
+    _code = "code"
+
+    code = fields.Char('Code')
+    comment = fields.Text('Comment')
+
+    _sql_constraints = [
+        ('attr_val_code_uniq', 'unique(code, attribute_id)',
+         "For each Attribute we must be found a unique 'code'"),
+        ]
