+++ product_code_builder/models/product_product.py
@@ -8,38 +8,6 @@ from openerp.exceptions import Warning as UserError
 from openerp.tools.translate import _
 
 
-class ProductTemplate(models.Model):
-    _inherit = "product.template"
-
-    def compute_default_auto_default_code(self):
-        if self._context.get('module') == 'product_code_builder':
-            # When we install the module we return False for the existing
-            # record
-            return False
-        return True
-
-    prefix_code = fields.Char(
-        string='Internal Reference',
-        help="This is the code of the product model"
-             "If Automatic Reference is checked, "
-             "this field is used as a prefix for "
-             "the product variant reference.\n"
-             "In case that there is only one variant "
-             "this code is the same as the code of the uniq variant")
-    default_code = fields.Char(related='prefix_code')
-    auto_default_code = fields.Boolean(
-        string='Automatic Reference',
-        default=compute_default_auto_default_code,
-        help="Generate a reference automatically "
-             "based on attribute codes")
-
-    _sql_constraints = [
-        ('uniq_prefix_code',
-         'unique(prefix_code)',
-         'The reference must be unique'),
-    ]
-
-
 class ProductProduct(models.Model):
     _inherit = "product.product"
