+++ product_code_builder/tests/__init__.py
@@ -1,1 +1,1 @@
-from . import test_code
+from . import test_product_code
