+++ setup/product_code_builder/odoo_addons/product_code_builder
@@ -1,0 +1,1 @@
+../../../product_code_builder
